+++ src/components/mainPage/InfoScreen/ConvenienceDescriptionItem.tsx
