+++ src/components/mainPage/InfoScreen/DriveDescriptionItem.tsx
