+++ src/components/mainPage/InfoScreen/GuideItem.tsx
