+++ src/components/mainPage/InfoScreen/PerformanceDescriptionItem.tsx
