+++ src/components/mainPage/MainScreen/EventHeader.tsx
