+++ src/pages/mainPage/InfoScreen/ColorSection.tsx
