+++ src/pages/mainPage/InfoScreen/InfoScreen.tsx
