+++ src/pages/mainPage/MainScreen/Timer.tsx
