+++ src/pages/mainPage/NavigationBar/NavElement.tsx
