+++ src/pages/mainPage/NavigationBar/NavigationBar.tsx
