+++ src/pages/mainPage/NotificationScreen/NotificationScreen.tsx
